Separated DroppedClasses into DroppedClasses and DroppedClass components
